Compute PGS from GWAS summary statistics using Bayesian sum of single-effect (SuSiE) linear regression using z scores

'rapidopgs_multi computes PGS from a from GWAS summary statistics using Bayesian sum of single-effect (SuSiE) linear regression using z scores

rapidopgs_multi(
  data,
  reference,
  ancestry = "EUR",
  pi_i = 1e-04,
  iterations = 100,
  ncores = 1,
  alpha.block = 1e-04,
  alpha.snp = 0.01,
  sd.prior = NULL
)

Arguments

data
a data.table containing GWAS summary statistic dataset with all required information.

reference
a string representing the path to the directory containing the reference panel (eg. "../ref-data/").

ancestry
a string indicating the ancestral population (DEFAULT: "EUR")

pi_i
a scalar representing the prior probability (DEFAULT: \(1 \times 10^{-4}\)).

iterations
number of maximum iterations at SuSie step. DEFAULT:100, which should be plenty to ensure convergence.

ncores
a numeric specifying the number of cores to be used. If not set, it won't use parallelisation.

alpha.block
a numeric threshold for minimum P-value in LD blocks. Blocks with minimum P above alpha.block will be skipped. Default: 1e-4.

Value

a data.table containing the sumstats dataset with computed PGS weights.

Details

Author

Guillermo Reales, Chris Wallace

Examples

if (FALSE) {
sumstats <- data.table(SNPID=c("rs139096444","rs3843766","rs61977545", "rs544733737",
    "rs2177641", "rs183491817", "rs72995775","rs78598863", "rs1411315"), 
    CHR=c("4","20","14","2","4","6","6","21","13"), 
    BP=c(1479959, 13000913, 29107209, 203573414, 57331393, 11003529, 149256398, 
        25630085, 79166661), 
    REF=c("C","C","C","T","G","C","C","G","T"), 
    ALT=c("A","T","T","A","A","A","T","A","C"), 
    ALT_FREQ=c(0.2611,0.4482,0.0321,0.0538,0.574,0.0174,0.0084,0.0304,0.7528),
    BETA=c(0.012,0.0079,0.0224,0.0033,0.0153,0.058,0.0742,0.001,-0.0131),
    SE=c(0.0099,0.0066,0.0203,0.0171,0.0063,0.0255,0.043,0.0188,0.0074),
    P=c(0.2237,0.2316,0.2682,0.8477,0.01473,0.02298,0.08472,0.9573,0.07535))
PGS  <- rapidopgs_multi(sumstats, reference = "ref-data/", ncores=2)
}