Posterior inclusion probabilities under a multiple causal variant model

A wrapper around susie_RSS from the susieR package. See https://stephenslab.github.io/susieR for the package and Wang et al. JRSSB 2020 for the paper describing the model https://rss.onlinelibrary.wiley.com/doi/full/10.1111/rssb.12388

susie_pip(d, LD, nref = 503, pi_i = 1e-04, sd.prior = NULL, max_it = 100)

Arguments

d
data.table with columns SNPID, BETA, SE (other columns allowed, d will be ignored)

LD
LD matrix with dimnames matching "SNPID" column

nref
number of individuals used to create the LD matrix

pi_i
prior probability that a given variant is causal. Default 1e-4

sd.prior
Standard deviation prior of the trait, if NULL (default), it will be estimated

max_it
maximum number of iterations for susie. The default of 100 should be plenty, but increase if there is not convergence

Value

returns pip for each SNP, which we will use as a snp weight in generating the PGS (=pip * BETA)

Details

NB: intended as an internal function, called from other functions in the RapidoPGS package.

Author

Chris Wallace, Guillermo Reales

Examples

data(michailidou) # load example data
d=michailidou[hm_chrom==3 & abs(hm_pos-27303612) < 1e+5] # focus on a window of association
setnames(d, old = c("hm_rsid", "hm_chrom", "hm_pos", "hm_other_allele",
  "hm_effect_allele", "hm_beta", "hm_effect_allele_frequency",
  "standard_error", "p_value"), new=c("SNPID","CHR", "BP","REF", "ALT",
  "BETA", "ALT_FREQ", "SE", "P")) # rename
LD=(1 - abs(outer(d$ALT_FREQ,d$ALT_FREQ, "-"))) *
   outer(sign(d$BETA), sign(d$BETA), "*")
dimnames(LD)=list(d$SNPID,d$SNPID)
susie_pip(d, LD)

#> Error in susie_pip(d, LD): could not find function "susie_pip"